Printer for hidden answer, condition for the end of the game
Adds a method for displaing a hidden answer in the way with "______" that shows correctly guessed letters. New loop which iterate until two conditions are fulfilled: the number of lives is greater than 0 and the game is not won ("areYouwinningSon" var). Adds arrays for guessed letters and actual number of lives.